+++ product-scenarios/3-securing-an-api-with-different-security-protocols/3.2-api-developers-want-to-secure-their-api-using-an-user-access-control-machanism/3.2.1-api-developer-wants-to-secure-api-resources-using-user-roles/src/test/java/org/wso2/am/scenario/tests/api/secure/userRoles/SecureUsingUserRolesNegativeTestCases.java
@@ -65,31 +65,10 @@ public class SecureUsingUserRolesNegativeTestCases extends ScenarioTestBase {
     private static final String ROLE_EXISTANCE = "isRoleExist";
     List<String> userList = new ArrayList();
     List<String> roleList = new ArrayList();
-    private String description = "This is a API creation description";
-    private String tag = "APICreationTag";
-    private String tierCollection = "Silver";
-    private String bizOwner = "wso2Test";
-    private String bizOwnerMail = "wso2test@gmail.com";
-    private String techOwner = "wso2";
-    private String techOwnerMail = "wso2@gmail.com";
-    private String endpointType = "secured";
-    private String endpointAuthType = "basicAuth";
-    private String epUsername = "wso2";
-    private String epPassword = "wso2123";
-    private String default_version_checked = "default_version";
-    private String responseCache = "enabled";
-    private String cacheTimeout = "300";
-    private String subscriptions = "all_tenants";
-    private String http_checked = "http";
-    private String https_checked = "";
-    private String inSequence = "debug_in_flow";
-    private String outSequence = "debug_out_flow";
     private String apiVersion = "1.0.0";
-    private String apiResource = "/find";
     private String apiVisibility = "public";
     private String backendEndPoint = "http://ws.cdyne.com/phoneverify/phoneverify.asmx";
-    private String apiName = "ScopeTest";
-    private String apiContext = "phone";
+    private String apiName = "APIScopeTestAPI";
     private File swagger_file;
     String resourceLocation = System.getProperty("test.resource.location");
 
@@ -132,33 +111,6 @@ public class SecureUsingUserRolesNegativeTestCases extends ScenarioTestBase {
         }
     }
 
-    public static String readFromFile(String file_name) throws IOException {
-        BufferedReader br = new BufferedReader(new FileReader(file_name));
-        StringBuilder sb = new StringBuilder();
-        int x;
-        while ((x = br.read()) != -1) {
-            sb.append((char) x);
-        }
-        String payloadText = sb.toString();
-        return payloadText;
-    }
-
-    private void createAndPublishAPI() throws APIManagerIntegrationTestException {
-        APIRequest apiRequest = new APIRequest(apiName, apiContext, apiVisibility, "", apiVersion, apiResource, description, tag,
-                tierCollection, backendEndPoint, bizOwner, bizOwnerMail, techOwner, techOwnerMail, endpointType,
-                endpointAuthType, epUsername, epPassword, default_version_checked, responseCache, cacheTimeout,
-                subscriptions, http_checked, https_checked, inSequence, outSequence);
-
-        //Design API with name,context,version,visibility,apiResource and with all optional values
-        HttpResponse serviceResponse = apiPublisher.addAPI(apiRequest);
-        verifyResponse(serviceResponse);
-        HttpResponse serviceResponseGetApi = apiPublisher.getAPI(apiName, SUPER_USER);
-
-        APILifeCycleStateRequest updateRequest = new APILifeCycleStateRequest(apiName, SUPER_USER,
-                APILifeCycleState.PUBLISHED);
-        apiPublisher.changeAPILifeCycleStatus(updateRequest);
-    }
-
     @DataProvider(name = "ScopeAndInValidRoleDataProvider")
     public static Object[][] ValidRoleDataProvider() {
         return new Object[][]{
@@ -175,7 +127,12 @@ public class SecureUsingUserRolesNegativeTestCases extends ScenarioTestBase {
         apiPublisher.login(SUPER_USER, SUPER_USER_LOGIN_PW);
         apiPublisherAdmin = new APIPublisherRestClient(publisherURL);
         apiPublisherAdmin.login(ADMIN_LOGIN_USERNAME, ADMIN_LOGIN_PW);
-        createAndPublishAPI();
+        try {
+            apiPublisher.developSampleAPI("swaggerFiles/apiScopeTest1.json",
+                    SUPER_USER, backendEndPoint, true, apiVisibility);
+        } catch (Exception ex) {
+            log.error("API publication failed", ex);
+        }
     }
 
     @Test(description = "3.2.1.9", dataProvider = "ScopeAndInValidRoleDataProvider",
@@ -191,7 +148,7 @@ public class SecureUsingUserRolesNegativeTestCases extends ScenarioTestBase {
     public void testScopeWithDuplicateKey() throws Exception {
         // This swagger will create "item_view" scope and assign it to a resource.
         swagger_file = new File(resourceLocation + File.separator + "swaggerFiles/APIScopeTest1.json");
-        String payload = readFromFile(swagger_file.getAbsolutePath());
+        String payload = Utils.readFromFile(swagger_file.getAbsolutePath());
         HttpResponse updateResponse = apiPublisher.updateResourceOfAPI(SUPER_USER, apiName, apiVersion, payload);
         verifyResponse(updateResponse);
         // Redeclare scope with item_view key.
